Simplify dependency and docker setup
We now declare the dependencies directly in the Pipfile
but do not install the reward_preprocessing package
itself via pipenv. This simplifies the Docker build and
in particular has the advantage that changes in the README,
setup.py, setup.cfg and pyproject.toml no longer require
rebuilding the dependencies stage.