Add dataset creation experiment
The new Experiment uses a trained model to create rollouts
and then stores those as (s, a, s', r) samples.
This dataset will be used to train a reward model in a
supervised manner.

I also restructured the tests slightly, using conftest.py
to define some fixtures that may be useful in future tests.