Implement reward model training
This is only a first prototype, I'm planning some refactoring
already. But this can serve as a working base to compare against
if issues occur when refactoring.

There is a bug in numpy (https://github.com/numpy/numpy/issues/18124)
that I avoid with a somewhat hacky workaround, see get_worker_init_fn
in datasets.py. Alternatives I considered was either just not allowing
num_workers > 0 or to use e.g. HDF5 instead of numpy to store results.
I think the workaround is preferable to both of those options, but it
would be even better to have one that doesn't duplicate code and is
less likely to lead to weird bugs down the line.