Fix determinism test
The old test could pass even in non-deterministic cases
because the untrained agents often got zero reward.
Now I've also added new tests which check that the
determinism test fails if different seeds are used
(as a sanity check).

The reason that seeding the environment isn't necessary
is that MiniGrid uses a fixed default seed. But we're still
seeding manually, first because we'll probably have other
envs in the future, and second to use different env seeds if
the global seed is different.