Refactor transitions
Transitions are now represented as instances of a new Transition
class. I wanted a common interface for transitions throughout the
codebase (rather than different conventions depending on whether we
have (s, a), (s, a, s'), just a state, etc.). A more lightweight
alternative to the Transition class would have been a namedtuple
or a dataclass. But I need to map functions over the fields in a
transition in several places (such as moving all tensors to a device),
which is why I wanted to have the Transition.apply method.