Improve rollout implementation
We now generate transitions using a dedicated function which we'll
be able to use in several more places in the future. It supports
vectorized environments with multiple environments inside them.

I've also taken this opportunity to simplify the shape of transition
and reward data everywhere. It previously had a dimension corresponding
to the environments in a vectorized environment. So far, this was just a
useless singleton dimension, and if we ever do use multiple environments,
it wouldn't have made sense at all anymore. I've removed this dimension
now.