Prohibit multiple workers with dynamic dataset
To support multiple workers, we would have to
ensure that each worker has its own copy of the
environment and that they are all seeded differently.
Should be doable but may be better to just use
stable-baseline's VecEnvs. In any case, we can
implement that if and when it's actually needed.