Small fix to to_torch
The previous version triggered a DeprecationWarning if
the input was a numpy scalar, such as numpy.bool_
The new version, which always uses torch.as_tensor,
avoids that and is also simpler (because we don't need
to distinguish between array/scalar, and we can convert
to float immediately).
Shouldn't change any functionality.