Make DynamicRewardData deterministic
As long as an agent was used, this was already
deterministic if used as part of the entire
pipeline, but this seems safer.

Without an agent, the previous version actually wasn't
deterministic, even as part of the reward model training
pipeline. Added a new test that would have caught that.