Switch reward models and trajectories to imitation
The original motivation was just to be able to import
and interpret RewardNets trained in imitation. Could have
also done that in some hacky way (and might have, if I'd
known how large this commit would become) but decided to
do it cleanly, which required also using imitation for
a lot of other stuff.